fix: refactoring of arborescence
