+++ mpqp/tools/circuit.py
@@ -7,8 +7,6 @@ import numpy.typing as npt
 from numpy.random import Generator
 
 from mpqp.core.circuit import QCircuit
-from mpqp.core.instruction.gates.custom_controlled_gate import CustomControlledGate
-from mpqp.core.instruction.gates.custom_gate import CustomGate
 from mpqp.core.instruction.gates.gate import Gate, SingleQubitGate
 from mpqp.core.instruction.gates.native_gates import (
     NATIVE_GATES,
@@ -24,7 +22,6 @@ from mpqp.core.instruction.gates.native_gates import (
     U,
 )
 from mpqp.core.instruction.gates.parametrized_gate import ParametrizedGate
-from mpqp.core.languages import Language
 from mpqp.noise.noise_model import (
     NOISE_MODELS,
     AmplitudeDamping,
@@ -37,8 +34,6 @@ from mpqp.noise.noise_model import (
 from mpqp.tools.maths import closest_unitary
 
 if TYPE_CHECKING:
-    from braket.circuits import Circuit as braket_Circuit
-    from cirq.circuits.circuit import Circuit as cirq_Circuit
     from qiskit.circuit import QuantumCircuit
     from qiskit._accelerate.circuit import CircuitInstruction
 
@@ -364,457 +359,3 @@ def replace_custom_gate(
         else:
             raise e
     return transpiled, transpiled.global_phase
-
-
-def _verify_convert_instructions(
-    gate: Gate, authorized_gates: set[type[Gate]], printing: bool = False
-) -> list[Gate]:
-    """Function used to verify if the instruction is contained in the gate set.
-    If the gate is a ComposedGate and not explicitly in the authorized_gates set it will check if it's decomposition is present in the gate set.
-    """
-    if len(authorized_gates) != 0:
-        if type(gate) not in authorized_gates:
-            raise ValueError(
-                f"The gate {type(gate)} are not in the set of authorized gates: f{authorized_gates}"
-            )
-        else:
-            return [gate]
-    if (
-        isinstance(gate, CustomControlledGate)
-        and isinstance(gate.non_controlled_gate, CustomGate)
-        and not printing
-    ):
-        # If the CustomControlledGate contains itself a custom gate it's better to return a bigger custom gate for compatibilities issues.
-        # If the non_controlled_gate is a NativeGate it shouldn't pose a problem.
-        # TODO: check how to decompose a ComposeGate that is inside a CCG, (example: a C-PRX)
-        return [gate.to_custom_gate()]
-
-    return [gate]
-
-
-def mpqp_to_qiskit(
-    circuit: QCircuit,
-    skip_pre_measure: bool = False,
-    skip_measurements: bool = False,
-    printing: bool = False,
-    authorized_gates: set[type[Gate]] | None = None,
-) -> QuantumCircuit:
-    """Translate a MPQP circuit to a Qiskit equivalent.
-
-    Note:
-        If the circuit contains ComposedGate type instructions you need to include the gate in the authorize_gates set for it to avoid decomposition, otherwise, the translation will always
-        try to use the gate's decomposition (see example).
-    Args:
-        circuit: The original MPQP circuit to be translated.
-        skip_pre_measure: If set at True will translate the circuit without its pre-measurement circuit (see QCircuit.to_other_language for more information).
-        skip_measurements: If set at True will translate the circuit without any measurement.
-        authorized_gates: The set of gates allowed on the circuit, if the circuit contains any other gates it raises a ValueError.
-
-    Examples:
-        >>> circuit = QCircuit([H(0), CNOT(0, 1), BasisMeasure()])
-        >>> cirq_with = mpqp_to_qiskit(circuit)
-        >>> cirq_without = mpqp_to_qiskit(circuit, skip_measurements=True)
-        >>> print(cirq_with) # doctest: +NORMALIZE_WHITESPACE
-             ┌───┐     ┌─┐
-        q_0: ┤ H ├──■──┤M├───
-             └───┘┌─┴─┐└╥┘┌─┐
-        q_1: ─────┤ X ├─╫─┤M├
-                  └───┘ ║ └╥┘
-        c: 2/═══════════╩══╩═
-                    0  1
-        >>> print(cirq_without) # doctest: +NORMALIZE_WHITESPACE
-             ┌───┐
-        q_0: ┤ H ├──■──
-             └───┘┌─┴─┐
-        q_1: ─────┤ X ├
-                  └───┘
-        c: 2/══════════
-    """
-    from qiskit.circuit import Operation, QuantumCircuit
-    from qiskit.circuit.quantumcircuit import CircuitInstruction
-    from qiskit.quantum_info import Operator
-
-    from mpqp.core.instruction import (
-        Measure,
-        Breakpoint,
-        CustomGate,
-        Barrier,
-        ControlledGate,
-        BasisMeasure,
-        ExpectationMeasure,
-    )
-
-    # to avoid defining twice the same parameter, we keep trace of the
-    # added parameters, and we use those instead of new ones when they
-    # are used more than once
-    if authorized_gates is None:
-        authorized_gates = set()
-    qiskit_parameters = set()
-    if circuit.nb_cbits == 0:
-        new_circ = QuantumCircuit(circuit.nb_qubits)
-    else:
-        new_circ = QuantumCircuit(circuit.nb_qubits, circuit.nb_cbits)
-
-    if circuit.label is not None:
-        new_circ.name = circuit.label
-
-    for instruction in circuit.instructions:
-        if isinstance(instruction, (Measure, Breakpoint)):
-            continue
-        options = {"printing": printing} if isinstance(instruction, CustomGate) else {}
-        if isinstance(instruction, Gate):
-            instr = _verify_convert_instructions(
-                instruction, authorized_gates, printing
-            )
-        else:
-            instr = [instruction]
-        for instruction in instr:
-            qiskit_inst = instruction.to_other_language(
-                Language.QISKIT, qiskit_parameters, **options
-            )
-            if TYPE_CHECKING:
-                assert isinstance(
-                    qiskit_inst, (CircuitInstruction, Operation, Operator)
-                )
-            cargs = []
-
-            if isinstance(instruction, CustomGate) and not isinstance(
-                instruction, CustomControlledGate
-            ):
-                if TYPE_CHECKING:
-                    assert isinstance(qiskit_inst, Operator)
-                if printing and len(instruction.free_symbols) > 0:
-                    new_circ.append(qiskit_inst, list(reversed(instruction.targets)))
-                else:
-                    new_circ.append(
-                        qiskit_inst,
-                        list(reversed(instruction.targets)),
-                    )
-            else:
-                qargs = []
-                if isinstance(instruction, ControlledGate):
-                    qargs = list(reversed(instruction.controls)) + list(
-                        reversed(instruction.targets)
-                    )
-                elif isinstance(instruction, Gate):
-                    qargs = list(reversed(instruction.targets))
-                elif isinstance(instruction, Barrier):
-                    qargs = range(circuit.nb_qubits)
-                else:
-                    raise ValueError(f"Instruction not handled: {instruction}")
-
-                if TYPE_CHECKING:
-                    assert not isinstance(qiskit_inst, Operator)
-                new_circ.append(
-                    qiskit_inst,
-                    list(qargs),
-                    cargs,
-                )
-    for measurement in circuit.measurements:
-        if not skip_pre_measure:
-
-            for pre_measure in measurement.pre_measure:
-                cargs = []
-                qiskit_pre_measure = pre_measure.to_other_language(
-                    Language.QISKIT, qiskit_parameters
-                )
-                new_circ.append(
-                    qiskit_pre_measure,
-                    list(reversed(pre_measure.targets)),
-                    cargs=cargs,
-                )
-        if not skip_measurements:
-            if isinstance(measurement, ExpectationMeasure):
-                continue
-            qiskit_inst = measurement.to_other_language(
-                Language.QISKIT, qiskit_parameters
-            )
-            if isinstance(measurement, BasisMeasure):
-                if TYPE_CHECKING:
-                    assert measurement.c_targets is not None
-            else:
-                raise ValueError(f"measurement not handled: {measurement}")
-
-            if TYPE_CHECKING:
-                assert not isinstance(qiskit_inst, Operator)
-            new_circ.append(
-                qiskit_inst,
-                [measurement.targets],
-                [measurement.c_targets],
-            )
-
-    new_circ.global_phase += (
-        circuit.input_g_phase
-        + circuit._generated_g_phase  # type: ignore[reportPrivateUsage]
-    )
-    return new_circ
-
-
-def mpqp_to_braket(
-    circuit: QCircuit,
-    skip_pre_measure: bool = False,
-    skip_measurements: bool = False,
-    authorized_gates: set[type[Gate]] | None = None,
-) -> braket_Circuit:
-    """Translate a MPQP circuit to a Braket equivalent.
-
-    Note:
-        If the circuit contains ComposedGate type instructions you need to include the gate in the authorize_gates set for it to avoid decomposition, otherwise, the translation will always
-        try to use the gate's decomposition (see example).
-    Args:
-        circuit: The original MPQP circuit to be translated.
-        skip_pre_measure: If set at True will translate the circuit without its pre-measurement circuit (see QCircuit.to_other_language for more information).
-        skip_measurements: If set at True will translate the circuit without any measurement.
-        authorized_gates: The set of gates allowed on the circuit, if the circuit contains any other gates it raises a ValueError.
-
-    Examples:
-        >>> circuit = QCircuit([H(0), CNOT(0, 1), BasisMeasure()])
-        >>> cirq_with = mpqp_to_braket(circuit)
-        >>> cirq_without = mpqp_to_braket(circuit, skip_measurements=True)
-        >>> print(cirq_with)  # doctest: +NORMALIZE_WHITESPACE
-        T  : │  0  │  1  │  2  │
-              ┌───┐       ┌───┐
-        q0 : ─┤ H ├───●───┤ M ├─
-              └───┘   │   └───┘
-                    ┌─┴─┐ ┌───┐
-        q1 : ───────┤ X ├─┤ M ├─
-                    └───┘ └───┘
-        T  : │  0  │  1  │  2  │
-        >>> print(cirq_without)  # doctest: +NORMALIZE_WHITESPACE
-        T  : │  0  │  1  │
-              ┌───┐
-        q0 : ─┤ H ├───●───
-              └───┘   │
-                    ┌─┴─┐
-        q1 : ───────┤ X ├─
-                    └───┘
-        T  : │  0  │  1  │
-    """
-    from mpqp.execution.providers.aws import apply_noise_to_braket_circuit
-    from mpqp.core.instruction import (
-        Measure,
-        Breakpoint,
-        Barrier,
-        ControlledGate,
-        BasisMeasure,
-    )
-
-    if authorized_gates is None:
-        authorized_gates = set()
-    if len(circuit.noises) != 0:
-        if any(isinstance(instr, CRk) for instr in circuit.instructions):
-            raise NotImplementedError(
-                "Cannot simulate noisy circuit with CRk gate due to "
-                "an error on AWS Braket side."
-            )
-    from braket.circuits import Circuit as BracketCircuit
-
-    braket_circuit = BracketCircuit()
-
-    # If the number of qubits are defined by the user, we ensure that every qubits are used.
-    # Otherwise the circuit can remain non continuous.
-    if circuit._user_nb_qubits is not None:  # pyright: ignore[reportPrivateUsage]
-        used_qubits = set().union(
-            *(
-                inst.connections()
-                for inst in circuit.instructions
-                if isinstance(inst, Gate)
-            )
-        )
-        if len(used_qubits) != circuit.nb_qubits:
-            from mpqp.gates import Id
-            from copy import deepcopy
-
-            circuit = QCircuit(
-                [
-                    Id(qubit)
-                    for qubit in range(circuit.nb_qubits)
-                    if qubit not in used_qubits
-                ],
-                nb_qubits=circuit.nb_qubits,
-            ) + deepcopy(circuit)
-
-    for instruction in circuit.instructions:
-        targets = [target for target in instruction.targets]
-        if isinstance(instruction, (Barrier, Breakpoint)):
-            continue
-        if isinstance(instruction, Measure):
-            if not skip_pre_measure:
-                for pre_measure in instruction.pre_measure:
-                    bracket_pre_measure = pre_measure.to_other_language(Language.BRAKET)
-                    braket_circuit.add(bracket_pre_measure, targets)
-            if not skip_measurements:
-                if isinstance(instruction, BasisMeasure) and instruction.shots != 0:
-                    braket_circuit.measure(targets)
-            continue
-        if isinstance(instruction, Gate):
-            instr = _verify_convert_instructions(instruction, authorized_gates)
-        else:
-            instr = [instruction]
-
-        for instruction in instr:
-            braket_instr = instruction.to_other_language(Language.BRAKET)
-            try:
-                targets = [target for target in instruction.targets]
-                if isinstance(instruction, CustomControlledGate):
-                    if isinstance(instruction.non_controlled_gate, CustomGate):
-                        targets = [
-                            control for control in instruction.controls
-                        ] + targets
-                        targets.sort()
-                elif isinstance(instruction, ControlledGate):
-                    targets = [control for control in instruction.controls] + targets
-                braket_circuit.add_instruction(braket_instr, target=targets)
-            except Exception as e:
-                raise ValueError(
-                    f"{type(braket_instr)}{braket_instr} cannot be added to the braket circuit: {e}"
-                )
-    if len(circuit.noises) != 0:
-        braket_circuit = apply_noise_to_braket_circuit(
-            braket_circuit,
-            circuit.noises,
-            circuit.nb_qubits,
-        )
-    return braket_circuit
-
-
-def mpqp_to_cirq(
-    circuit: QCircuit,
-    skip_pre_measure: bool = False,
-    skip_measurements: bool = False,
-    authorized_gates: set[type[Gate]] | None = None,
-) -> cirq_Circuit:
-    """Translate a MPQP circuit to a Cirq equivalent.
-
-    Note:
-        If the circuit contains ComposedGate type instructions you need to include the gate in the authorize_gates set for it to avoid decomposition, otherwise, the translation will always
-        try to use the gate's decomposition (see example). Also in the case of CustomGates it will always be decomposed into the {Ry, Rz, CNOT} basis for translation compatibilities.
-    Args:
-        circuit: The original MPQP circuit to be translated.
-        skip_pre_measure: If set at True will translate the circuit without its pre-measurement circuit (see QCircuit.to_other_language for more information).
-        skip_measurements: If set at True will translate the circuit without any measurement.
-        authorized_gates: The set of gates allowed on the circuit, if the circuit contains any other gates it raises a ValueError.
-
-    Examples:
-        >>> circuit = QCircuit([H(0), CNOT(0,1), BasisMeasure()])
-        >>> cirq_with = mpqp_to_cirq(circuit)
-        >>> cirq_without = mpqp_to_cirq(circuit, skip_measurements=True)
-        >>> print(cirq_with) # doctest: +NORMALIZE_WHITESPACE
-        q_0: ───I───H───@───M('')───
-                        │   │
-        q_1: ───I───────X───M───────
-        >>> print(cirq_without) # doctest: +NORMALIZE_WHITESPACE
-        q_0: ───I───H───@───
-                        │
-        q_1: ───I───────X───
-    """
-    # TODO: add better doc for all providers (ComposedGate) and CustomGate behavior for this one specifically
-    from cirq.circuits.circuit import Circuit as CirqCircuit
-    from cirq.ops.identity import I
-    from cirq.ops.named_qubit import NamedQubit
-    from mpqp.core.instruction import (
-        Measure,
-        Breakpoint,
-        CustomGate,
-        Barrier,
-        ControlledGate,
-        CustomControlledGate,
-        ExpectationMeasure,
-    )
-
-    if authorized_gates is None:
-        authorized_gates = set()
-    cirq_qubits = [NamedQubit(f"q_{i}") for i in range(circuit.nb_qubits)]
-    cirq_circuit = CirqCircuit()
-
-    for qubit in cirq_qubits:
-        cirq_circuit.append(I(qubit))
-
-    for instruction in circuit.instructions:
-        if not skip_pre_measure:
-            if isinstance(instruction, Measure):
-                for pre_measure in instruction.pre_measure:
-                    if isinstance(pre_measure, (CustomGate, CustomControlledGate)):
-                        instr = _verify_convert_instructions(
-                            pre_measure, authorized_gates
-                        )
-                        qasm2_code, gphase = pre_measure.to_other_language(
-                            Language.QASM2
-                        )  # pyright: ignore[reportGeneralTypeIssues]
-                        if TYPE_CHECKING:
-                            assert isinstance(qasm2_code, str)
-                        from mpqp.qasm.qasm_to_cirq import qasm2_to_cirq_Circuit
-
-                        qasm2_code = (
-                            "OPENQASM 2.0;"
-                            + "\ninclude \"qelib1.inc\";"
-                            + f"\nqreg q[{circuit.nb_qubits}];\n"
-                            + qasm2_code
-                        )
-                        custom_cirq_circuit = qasm2_to_cirq_Circuit(qasm2_code)
-                        cirq_circuit += custom_cirq_circuit
-                        # TODO: handle gphase in the circuit
-                        circuit._generated_g_phase += gphase  # type: ignore[reporPrivateUsage]
-                    else:
-                        cirq_pre_measure = pre_measure.to_other_language(Language.CIRQ)
-                        targets = []
-                        for target in pre_measure.targets:
-                            targets.append(cirq_qubits[target])
-                        cirq_circuit.append(cirq_pre_measure.on(*targets))
-
-        if isinstance(instruction, Gate):
-            instr = _verify_convert_instructions(instruction, authorized_gates)
-        else:
-            instr = [instruction]
-
-        for gate in instr:
-            if isinstance(gate, (ExpectationMeasure, Barrier, Breakpoint)):
-                continue
-            elif isinstance(gate, CustomGate):
-                from cirq.ops.raw_types import Gate as CirqGate
-
-                custom_gate = instr[0]
-
-                targets = []
-                for target in custom_gate.targets:
-                    targets.append(cirq_qubits[target])
-
-                cirq_instruction = custom_gate.to_other_language(Language.CIRQ)
-                assert isinstance(cirq_instruction, CirqGate)
-
-                cirq_circuit.append(cirq_instruction.on(*targets))
-                continue
-            elif isinstance(gate, ControlledGate):
-                targets = []
-                for target in gate.targets:
-                    targets.append(cirq_qubits[target])
-                controls = []
-                for control in gate.controls:
-                    controls.append(cirq_qubits[control])
-                cirq_instruction = gate.to_other_language(Language.CIRQ)
-                cirq_circuit.append(cirq_instruction.on(*controls, *targets))
-            else:
-                if skip_measurements and isinstance(gate, Measure):
-                    continue
-                targets = []
-                for target in gate.targets:
-                    targets.append(cirq_qubits[target])
-                cirq_instruction = gate.to_other_language(Language.CIRQ)
-                if TYPE_CHECKING:
-                    assert cirq_instruction
-                cirq_circuit.append(cirq_instruction.on(*targets))
-
-    if circuit.noises:
-        from mpqp.execution.providers.google import apply_noise_to_cirq_circuit
-
-        return apply_noise_to_cirq_circuit(
-            cirq_circuit,
-            circuit.noises,
-        )
-
-    if circuit.input_g_phase != 0:
-        from cirq import GlobalPhaseGate
-
-        cirq_circuit.insert(0, GlobalPhaseGate(np.exp(1j * circuit.input_g_phase)).on())
-
-    return cirq_circuit
