+++ tests/qasm/test_qasm_to_qiskit.py
@@ -1,6 +1,6 @@
 import pytest
 
-from mpqp.qasm.qasm_to_qiskit import qasm2_to_Qiskit_Circuit
+from mpqp.translation.qasm.qasm_to_qiskit import qasm2_to_Qiskit_Circuit
 
 
 @pytest.mark.provider("qiskit")
